requirements.txt
Flask-RESTful==0.3.6
marshmallow==3.0.0b13
mock==2.0.0
pytest==3.7.4
requests==2.19.1
vcrpy==1.13.0